serial: ns16550: add config to use readl/writel
If either of CONFIG_SYS_LITTLE_ENDIAN or CONFIG_SYS_BIG_ENDIAN
are set then the code tries to use the corresponding IO endian
which may not be right.

Add a config CONFIG_SYS_NS16550_MEM32_MMIO to allow selecting
the default readl/writel functions and allow those to deal with
the endian.

Signed-off-by: Ben Dooks <ben.dooks@codethink.co.uk>